feat(runtime): enable distributed execution.

+++ compiler/include/concretelang/Runtime/dfr_debug_interface.h
@@ -12,7 +12,7 @@
 extern "C" {
 size_t _dfr_debug_get_node_id();
 size_t _dfr_debug_get_worker_id();
-void _dfr_debug_print_task(const char *name, int inputs, int outputs);
+void _dfr_debug_print_task(const char *name, size_t inputs, size_t outputs);
 void _dfr_print_debug(size_t val);
 }
 #endif
